requirements.1.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
cryptography==2.1.4
Django==1.11
django-push-notifications==1.6.0
django-tastypie==0.14.0
djangorestframework==3.7.7
ecdsa==0.13
enum34==1.1.6
Flask==0.12.2
h2==2.6.2
hpack==3.0.0
http-ece==1.0.5
hyper==0.7.0
hyperframe==3.2.0
idna==2.6
ipaddress==1.0.19
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.12
Pillow==5.0.0
py-vapid==1.3.0
pycparser==2.18
PyJWT==1.5.3
PyMySQL==0.8.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2017.3
pywebpush==1.5.0
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
